move settings to seperate file
