+++ tools/cabana/mainwin.h
@@ -16,7 +16,7 @@ public:
 
 protected:
   void closeEvent(QCloseEvent *event) override;
-  void setOption();
+  void openSettingsDlg();
   void saveSession();
   void restoreSession();
 
@@ -27,16 +27,4 @@ protected:
   QSplitter *splitter;
   QWidget *floating_window = nullptr;
   QVBoxLayout *r_layout;
-};
-
-class SettingsDlg : public QDialog {
-  Q_OBJECT
-
-public:
-  SettingsDlg(QWidget *parent);
-  void save();
-  QSpinBox *fps;
-  QSpinBox *log_size ;
-  QSpinBox *cached_segment;
-  QSpinBox *chart_height;
 };
